Move contents of RPCConeBuilder.h to the source
